Integrated level builder with game
BoxTerm is now a host for the menu bar. Game functionality has moved to
SokobanGame. You can switch between modes within the same window and the
current map is inherited (including resetting objects to starting
positions when going from game to editor). The editor has inherited some
functionality from BoxTerm because of this.