Renamed SpriteMap.getMap() to getSokobanMap()
The old name was slightly ambiguous. I have also added a method to
SokobanGame and LevelBuilder to make the active SokobanMap directly
accessible so you don't need to keep calling
SomeClass.getSpriteMap().getSokobanMap().

When we get to the polish stage I think I'll make a new class to contain
the code that's duplicated between SokobanGame and LevelBuilder and have
those two classes extend the new one.