Updated level editor GUI
The palette is now placed horizontally at the bottom of the window.
I did this to try and prevent the window from resizing so much when
switching modes. That looks weird to me and I think it's not something
you see very often in programs so I want to minimise it. This solution
isn't quite perfect as the height of the status bar and tile palette
aren't the same and I'm struggling to force them to be the same size.
Also, whenever the magnitude of the box/goal counters increases, the
palette gets shifted to the side a little.