Fixed calculation of gCost in path finder
Previously, the path finder gave an equal "distance from player" penalty
to all nodes. Somehow in conjunction with me accidentally sorting the
nodes by most expensive first this produced a pretty good pathfinder,
but I've fixed that oversight now.

Basically, the path finder should now better handle walls and boxes.

There's a block of code for making sure that each visited node belongs
to a parent that gives it the lowest gCost. I'm not actually sure that
this is required in a game using my selected heuristic. I've added a
"tell Tom" print statement to that block of code so please tell me if
you ever see it crop up.